# frontend/openapi/clinica.json
{
  "openapi": "3.0.3",
  "info": { "title": "ClinicaSaaS", "version": "1.0.0" },
  "paths": {
    "/api/plataforma/acceso/ingreso": {
      "post": {
        "operationId": "platformLogin",
        "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LoginCommand" } } } },
        "responses": { "200": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TokenDto" } } } } }
      }
    },
    "/api/plataforma/consultorios": {
      "get": {
        "operationId": "listTenants",
        "responses": { "200": { "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/TenantDto" } } } } } }
      },
      "post": {
        "operationId": "createTenant",
        "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateTenantCommand" } } } },
        "responses": { "200": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TenantDto" } } } } }
      }
    },
    "/api/plataforma/consultorios/{id}/estado": {
      "patch": {
        "operationId": "changeTenantStatus",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ChangeTenantStatusCommand" } } } },
        "responses": { "200": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TenantDto" } } } } }
      }
    },
    "/api/acceso/ingreso": {
      "post": {
        "operationId": "tenantLogin",
        "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LoginCommand" } } } },
        "responses": { "200": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TokenDto" } } } } }
      }
    },
    "/api/acceso/registro": {
      "post": {
        "operationId": "registerPatient",
        "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RegisterPatientCommand" } } } },
        "responses": { "200": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TokenDto" } } } } }
      }
    },
    "/api/sedes": {
      "get": { "operationId": "locations", "responses": { "200": { "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/LocationDto" } } } } } } },
      "post": {
        "operationId": "createLocation",
        "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SaveLocationCommand" } } } },
        "responses": { "200": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LocationDto" } } } } }
      }
    },
    "/api/sedes/{id}": {
      "put": {
        "operationId": "updateLocation",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SaveLocationCommand" } } } },
        "responses": { "200": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LocationDto" } } } } }
      }
    },
    "/api/servicios": {
      "get": { "operationId": "medicalServices", "responses": { "200": { "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/MedicalServiceDto" } } } } } } },
      "post": {
        "operationId": "createMedicalService",
        "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SaveMedicalServiceCommand" } } } },
        "responses": { "200": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MedicalServiceDto" } } } } }
      }
    },
    "/api/especialidades": {
      "get": { "operationId": "specialties", "responses": { "200": { "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/SpecialtyDto" } } } } } } },
      "post": {
        "operationId": "createSpecialty",
        "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SaveSpecialtyCommand" } } } },
        "responses": { "200": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SpecialtyDto" } } } } }
      }
    },
    "/api/tipos-turno": {
      "get": { "operationId": "appointmentTypes", "responses": { "200": { "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/AppointmentTypeDto" } } } } } } },
      "post": {
        "operationId": "createAppointmentType",
        "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SaveAppointmentTypeCommand" } } } },
        "responses": { "200": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AppointmentTypeDto" } } } } }
      }
    },
    "/api/profesionales": {
      "get": {
        "operationId": "professionals",
        "parameters": [{ "name": "specialtyId", "in": "query", "schema": { "type": "string" } }],
        "responses": { "200": { "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/ProfessionalDto" } } } } } }
      }
    },
    "/api/empleados/importar": {
      "post": {
        "operationId": "importEmployees",
        "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ImportEmployeesCommand" } } } },
        "responses": { "200": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EmployeeImportDto" } } } } }
      }
    },
    "/api/agendas": {
      "get": {
        "operationId": "getSchedule",
        "parameters": [{ "name": "professionalId", "in": "query", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/ScheduleBlockDto" } } } } } }
      },
      "put": {
        "operationId": "saveSchedule",
        "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SaveScheduleCommand" } } } },
        "responses": { "200": { "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/ScheduleBlockDto" } } } } } }
      }
    },
    "/api/bloqueos": {
      "get": {
        "operationId": "blockouts",
        "parameters": [{ "name": "professionalId", "in": "query", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/BlockoutDto" } } } } } }
      },
      "post": {
        "operationId": "createBlockout",
        "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateBlockoutCommand" } } } },
        "responses": { "200": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BlockoutDto" } } } } }
      }
    },
    "/api/bloqueos/{id}": {
      "delete": {
        "operationId": "deleteBlockout",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "204": { "description": "Deleted" } }
      }
    },
    "/api/disponibilidad": {
      "get": {
        "operationId": "availability",
        "parameters": [
          { "name": "locationId", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "specialtyId", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "professionalId", "in": "query", "schema": { "type": "string" } },
          { "name": "date", "in": "query", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/SlotDto" } } } } } }
      }
    },
    "/api/turnos": {
      "get": {
        "operationId": "appointmentsForDay",
        "parameters": [
          { "name": "date", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "professionalId", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/AppointmentDto" } } } } } }
      },
      "post": {
        "operationId": "bookAppointment",
        "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BookAppointmentCommand" } } } },
        "responses": { "200": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AppointmentDto" } } } } }
      }
    },
...
    "/api/turnos/{id}/cancelar": {
      "post": {
        "operationId": "cancelAppointment",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AppointmentDto" } } } } }
      }
    },
    "/api/turnos/{id}/admitir": {
      "post": {
        "operationId": "checkInAppointment",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AppointmentDto" } } } } }
      }
    },
    "/api/turnos/{id}/reprogramar": {
      "post": {
        "operationId": "rescheduleAppointment",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RescheduleAppointmentCommand" } } } },
        "responses": { "200": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AppointmentDto" } } } } }
      }
    },
    "/api/lista-espera": {
      "get": {
        "operationId": "waitlist",
        "responses": { "200": { "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/WaitlistEntryDto" } } } } } }
      },
      "post": {
        "operationId": "joinWaitlist",
        "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateWaitlistEntryCommand" } } } },
        "responses": { "200": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/WaitlistEntryDto" } } } } }
      }
    },
    "/api/lista-espera/{id}/asignar": {
      "post": {
        "operationId": "assignWaitlistEntry",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AssignWaitlistEntryCommand" } } } },
        "responses": { "200": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AppointmentDto" } } } } }
      }
    },
...
    "/api/encuentros": {
      "post": {
        "operationId": "saveEncounter",
        "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SaveEncounterCommand" } } } },
        "responses": { "200": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EncounterDto" } } } } }
      }
    },
    "/api/encuentros/{id}/cerrar": {
      "post": {
        "operationId": "closeEncounter",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EncounterDto" } } } } }
      }
    },
    "/api/recetas": {
      "post": {
        "operationId": "createPrescription",
        "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreatePrescriptionCommand" } } } },
        "responses": { "200": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PrescriptionDto" } } } } }
      }
    },
...
    "/api/recetas/{id}": {
      "get": {
        "operationId": "getPrescription",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PrescriptionDto" } } } } }
      }
    },
    "/api/historia": {
      "get": {
        "operationId": "clinicalHistory",
        "parameters": [{ "name": "patientId", "in": "query", "schema": { "type": "string" } }],
        "responses": { "200": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ClinicalHistoryDto" } } } } }
      }
    },
...
    "/api/historias-medicas/{patientId}": {
      "get": {
        "operationId": "medicalRecord",
        "parameters": [{ "name": "patientId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MedicalRecordDto" } } } } }
      },
      "put": {
        "operationId": "saveMedicalRecord",
        "parameters": [{ "name": "patientId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SaveMedicalRecordCommand" } } } },
        "responses": { "200": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MedicalRecordDto" } } } } }
      }
    },
    "/api/aranceles": {
      "get": {
        "operationId": "fees",
        "responses": { "200": { "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/FeeDto" } } } } } }
      },
      "post": {
        "operationId": "createFee",
        "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateFeeCommand" } } } },
        "responses": { "200": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/FeeDto" } } } } }
      }
    },
...
    "/api/comprobantes/{id}/pagar": {
      "post": {
        "operationId": "payInvoice",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PayInvoiceCommand" } } } },
        "responses": { "200": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/InvoiceDto" } } } } }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "LoginCommand": { "type": "object", "required": ["email", "password"], "properties": { "email": { "type": "string" }, "password": { "type": "string" } } },
      "TokenDto": { "type": "object", "required": ["token", "email", "name", "role", "mustChangePassword"], "properties": { "token": { "type": "string" }, "email": { "type": "string" }, "name": { "type": "string" }, "role": { "type": "string" }, "tenantSlug": { "type": "string", "nullable": true }, "mustChangePassword": { "type": "boolean" } } },
      "CreateTenantCommand": { "type": "object", "required": ["slug", "name", "type"], "properties": { "slug": { "type": "string" }, "name": { "type": "string" }, "type": { "$ref": "#/components/schemas/TenantType" } } },
      "ChangeTenantStatusCommand": { "type": "object", "required": ["status"], "properties": { "status": { "$ref": "#/components/schemas/TenantStatus" } } },
      "TenantDto": { "type": "object", "required": ["id", "slug", "name", "type", "status", "createdAt"], "properties": { "id": { "type": "string" }, "slug": { "type": "string" }, "name": { "type": "string" }, "type": { "$ref": "#/components/schemas/TenantType" }, "status": { "$ref": "#/components/schemas/TenantStatus" }, "createdAt": { "type": "string" } } },
      "RegisterPatientCommand": { "type": "object", "required": ["email", "password", "firstName", "lastName", "documentNumber", "birthDate", "phone"], "properties": { "email": { "type": "string" }, "password": { "type": "string" }, "firstName": { "type": "string" }, "lastName": { "type": "string" }, "documentNumber": { "type": "string" }, "birthDate": { "type": "string" }, "phone": { "type": "string" } } },
      "LocationDto": { "type": "object", "required": ["id", "name", "address", "isActive"], "properties": { "id": { "type": "string" }, "name": { "type": "string" }, "address": { "type": "string" }, "isActive": { "type": "boolean" } } },
      "SaveLocationCommand": { "type": "object", "required": ["name", "address"], "properties": { "name": { "type": "string" }, "address": { "type": "string" } } },
      "MedicalServiceDto": { "type": "object", "required": ["id", "locationId", "name"], "properties": { "id": { "type": "string" }, "locationId": { "type": "string" }, "name": { "type": "string" } } },
      "SaveMedicalServiceCommand": { "type": "object", "required": ["locationId", "name"], "properties": { "locationId": { "type": "string" }, "name": { "type": "string" } } },
      "SpecialtyDto": { "type": "object", "required": ["id", "name"], "properties": { "id": { "type": "string" }, "name": { "type": "string" } } },
      "SaveSpecialtyCommand": { "type": "object", "required": ["name"], "properties": { "name": { "type": "string" } } },
      "AppointmentTypeDto": { "type": "object", "required": ["id", "name", "durationMinutes"], "properties": { "id": { "type": "string" }, "name": { "type": "string" }, "durationMinutes": { "type": "integer" }, "specialtyId": { "type": "string", "nullable": true } } },
      "SaveAppointmentTypeCommand": { "type": "object", "required": ["name", "durationMinutes"], "properties": { "name": { "type": "string" }, "durationMinutes": { "type": "integer" }, "specialtyId": { "type": "string", "nullable": true } } },
      "ProfessionalDto": { "type": "object", "required": ["id", "firstName", "lastName", "email"], "properties": { "id": { "type": "string" }, "firstName": { "type": "string" }, "lastName": { "type": "string" }, "email": { "type": "string" }, "licenseNumber": { "type": "string", "nullable": true }, "specialtyId": { "type": "string", "nullable": true } } },
      "ImportEmployeesCommand": { "type": "object", "required": ["csv"], "properties": { "csv": { "type": "string" } } },
...
      "EmployeeImportDto": { "type": "object", "required": ["created", "rejected"], "properties": { "created": { "type": "array", "items": { "$ref": "#/components/schemas/CreatedEmployeeDto" } }, "rejected": { "type": "array", "items": { "$ref": "#/components/schemas/RejectedRowDto" } } } },
      "ScheduleBlockDto": { "type": "object", "required": ["id", "day", "startTime", "endTime", "locationId", "appointmentTypeId"], "properties": { "id": { "type": "string" }, "day": { "$ref": "#/components/schemas/DayOfWeek" }, "startTime": { "type": "string" }, "endTime": { "type": "string" }, "locationId": { "type": "string" }, "appointmentTypeId": { "type": "string" } } },
      "SaveScheduleCommand": { "type": "object", "required": ["professionalId", "blocks"], "properties": { "professionalId": { "type": "string" }, "blocks": { "type": "array", "items": { "$ref": "#/components/schemas/ScheduleBlockDto" } } } },
      "BlockoutDto": { "type": "object", "required": ["id", "professionalId", "start", "end", "reason"], "properties": { "id": { "type": "string" }, "professionalId": { "type": "string" }, "start": { "type": "string" }, "end": { "type": "string" }, "reason": { "type": "string" } } },
      "CreateBlockoutCommand": { "type": "object", "required": ["professionalId", "start", "end", "reason"], "properties": { "professionalId": { "type": "string" }, "start": { "type": "string" }, "end": { "type": "string" }, "reason": { "type": "string" } } },
      "SlotDto": { "type": "object", "required": ["professionalId", "professional", "locationId", "appointmentTypeId", "appointmentType", "start", "end"], "properties": { "professionalId": { "type": "string" }, "professional": { "type": "string" }, "locationId": { "type": "string" }, "appointmentTypeId": { "type": "string" }, "appointmentType": { "type": "string" }, "start": { "type": "string" }, "end": { "type": "string" } } },
      "BookAppointmentCommand": { "type": "object", "required": ["professionalId", "locationId", "appointmentTypeId", "start"], "properties": { "patientId": { "type": "string", "nullable": true }, "professionalId": { "type": "string" }, "locationId": { "type": "string" }, "appointmentTypeId": { "type": "string" }, "start": { "type": "string" }, "visitReason": { "type": "string", "nullable": true } } },
      "AppointmentDto": { "type": "object", "required": ["id", "patientId", "patient", "professionalId", "professional", "locationId", "location", "appointmentTypeId", "appointmentType", "start", "end", "status"], "properties": { "id": { "type": "string" }, "patientId": { "type": "string" }, "patient": { "type": "string" }, "professionalId": { "type": "string" }, "professional": { "type": "string" }, "locationId": { "type": "string" }, "location": { "type": "string" }, "appointmentTypeId": { "type": "string" }, "appointmentType": { "type": "string" }, "start": { "type": "string" }, "end": { "type": "string" }, "status": { "$ref": "#/components/schemas/AppointmentStatus" }, "visitReason": { "type": "string", "nullable": true } } },
      "RescheduleAppointmentCommand": { "type": "object", "required": ["start"], "properties": { "start": { "type": "string" } } },
      "WaitlistEntryDto": { "type": "object", "required": ["id", "patientId", "patient", "locationId", "specialtyId", "status", "createdAt"], "properties": { "id": { "type": "string" }, "patientId": { "type": "string" }, "patient": { "type": "string" }, "professionalId": { "type": "string", "nullable": true }, "locationId": { "type": "string" }, "specialtyId": { "type": "string" }, "status": { "$ref": "#/components/schemas/WaitlistStatus" }, "createdAt": { "type": "string" }, "notes": { "type": "string", "nullable": true } } },
      "CreateWaitlistEntryCommand": { "type": "object", "required": ["locationId", "specialtyId"], "properties": { "patientId": { "type": "string", "nullable": true }, "professionalId": { "type": "string", "nullable": true }, "locationId": { "type": "string" }, "specialtyId": { "type": "string" }, "notes": { "type": "string", "nullable": true } } },
      "AssignWaitlistEntryCommand": { "type": "object", "required": ["start", "appointmentTypeId"], "properties": { "start": { "type": "string" }, "appointmentTypeId": { "type": "string" } } },
...
      "CreatePrescriptionCommand": { "type": "object", "required": ["encounterId", "items"], "properties": { "encounterId": { "type": "string" }, "instructions": { "type": "string", "nullable": true }, "items": { "type": "array", "items": { "$ref": "#/components/schemas/PrescriptionItemDto" } } } },
...
      "PayInvoiceCommand": { "type": "object", "required": ["method"], "properties": { "method": { "$ref": "#/components/schemas/PaymentMethod" } } },
...
    }
  }
}